\name{estjointpab-package}
\alias{estjointpab-package}
\alias{estjointpab}
\docType{package}
\title{
\packageTitle{estjointpab}
}
\description{
\packageDescription{estjointpab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{estjointpab}
\packageIndices{estjointpab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{estjointpab}

Maintainer: \packageMaintainer{estjointpab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
